refactor: Move links to dedicated folder
