+++ packages/cozy-client/src/index.node.js
@@ -1,6 +1,6 @@
 export { default } from './CozyClient'
-export { default as CozyLink } from './CozyLink'
-export { default as StackLink } from './StackLink'
+export { default as CozyLink } from './links/CozyLink'
+export { default as StackLink } from './links/StackLink'
 export { default as compose } from 'lodash/flow'
 export {
   QueryDefinition,
