+++ packages/cozy-client/src/links/WebFlagshipLink.js
