+++ packages/cozy-client/types/index.d.ts
@@ -1,7 +1,8 @@
 export { default } from "./CozyClient";
-export { default as CozyLink } from "./CozyLink";
-export { default as StackLink } from "./StackLink";
-export { default as WebFlagshipLink } from "./WebFlagshipLink";
+export { default as CozyLink } from "./links/CozyLink";
+export { default as StackLink } from "./links/StackLink";
+export { default as WebFlagshipLink } from "./links/WebFlagshipLink";
+export { default as DataProxyLink } from "./links/DataProxyLink";
 export { default as compose } from "lodash/flow";
 export { default as Registry } from "./registry";
 export { default as RealTimeQueries } from "./RealTimeQueries";
