+++ packages/cozy-client/types/index.node.d.ts
@@ -1,6 +1,6 @@
 export { default } from "./CozyClient";
-export { default as CozyLink } from "./CozyLink";
-export { default as StackLink } from "./StackLink";
+export { default as CozyLink } from "./links/CozyLink";
+export { default as StackLink } from "./links/StackLink";
 export { default as compose } from "lodash/flow";
 export { getReferencedBy } from "./associations/helpers";
 export { cancelable } from "./utils";
