+++ packages/cozy-client/types/links/CozyLink.d.ts
