+++ packages/cozy-client/types/links/WebFlagshipLink.d.ts
